requirements.txt
plexapi==4.15.13
pyarr==5.2.0
python-dotenv==1.0.1
jellyfin-apiclient-python==1.11.0